Remove obsolete, wrong, and unreachable IDN wildcard logic.
mozilla::pkix enforced the rule the comment describes at one point.
However, it later dropped support for wildcards of the form
"foo*.example.org" where the "*" isn't the only character in the
wildcard. webpki implements that newer logic. With that newer logic,
it wasn't possible for `is_wildcard` to be set and for the hostname
to start with "xn--" since `is_wildcard` is only set when the hostname
starts with "*.".