Better error in verify_is_valid_tls_server_cert
Before this commit `UnknownIssuer` was emitted on any error of
verifying the anchor to the certificate.

Now the most important error is returned. In particular,
`UnsupportedSignatureAlgorithm` is returned on, well, unsupported
signature algorithm.

I agree to license my contributions to each file under the terms
given at the top of each file I changed.